Dijkstra et A* marchent parfaitement

+++ be-graphes-algos/src/main/java/org/insa/graphs/algorithm/shortestpath/LabelStar.java
@@ -4,7 +4,7 @@ import org.insa.graphs.model.Node;
 public class LabelStar extends Label {
 	
 	// Valeur courante du plus court chemin depuis l'origine vers le sommet
-	public float estim;
+	public double estim;
 
 	public LabelStar(Node courant) {
 		super(courant);
@@ -12,12 +12,7 @@ public class LabelStar extends Label {
 	}
 	
 	@Override
-	public float getScore() {
+	public double getScore() {
 		return this.cout + this.estim;
 	}
-	
-//	public float minLabel(LabelStar other, float val) {
-//		System.out.println("A*");
-//		return Math.min(this.getScore(), other.getCost() + other.estim + val);
-//    }
 }
